Refactor: admin 폴더를 삭제하고 auth 폴더로 옮긴다
