+++ server/src/main/java/com/ryc/api/v2/auth/infra/jpa/AdminJpaRepository.java
@@ -1,10 +1,10 @@
-package com.ryc.api.v2.admin.infra.jpa;
+package com.ryc.api.v2.auth.infra.jpa;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ryc.api.v2.admin.infra.entity.AdminEntity;
+import com.ryc.api.v2.auth.infra.entity.AdminEntity;
 
 public interface AdminJpaRepository extends JpaRepository<AdminEntity, String> {
   boolean existsByEmail(String email);
