+++ server/src/main/java/com/ryc/api/v2/auth/presentation/request/RegisterRequest.java
@@ -2,7 +2,7 @@ package com.ryc.api.v2.auth.presentation.request;
 
 import jakarta.validation.constraints.NotEmpty;
 
-import com.ryc.api.v2.admin.domain.AdminDefaultRole;
+import com.ryc.api.v2.auth.domain.AdminDefaultRole;
 
 public record RegisterRequest(
     @NotEmpty(message = "name is empty") String name,
