+++ server/src/main/java/com/ryc/api/v2/club/domain/ClubRepository.java
@@ -3,6 +3,8 @@ package com.ryc.api.v2.club.domain;
 import java.util.List;
 import java.util.Optional;
 
+import com.ryc.api.v2.auth.domain.Admin;
+
 public interface ClubRepository {
   Club save(Club club);
 
@@ -11,4 +13,6 @@ public interface ClubRepository {
   boolean existsByName(String name);
 
   List<Club> findAll();
+
+  Role assignOwer(Club club, Admin admin);
 }
