+++ server/src/main/java/com/ryc/api/v2/club/presentation/ClubHttpApi.java
@@ -32,6 +32,8 @@ public class ClubHttpApi {
   @PostMapping
   @Operation(summary = "동아리 생성 API")
   public ResponseEntity<ClubCreateResponse> createClub(@Valid @RequestBody ClubCreateRequest body) {
+    System.out.println("==============================");
+
     ClubCreateResponse response = clubService.createClub(body);
     return ResponseEntity.status(HttpStatus.CREATED).body(response);
   }
