+++ server/src/main/java/com/ryc/api/v2/security/service/CustomUserDetailService.java
@@ -6,8 +6,8 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import com.ryc.api.v2.admin.domain.Admin;
-import com.ryc.api.v2.admin.domain.AdminRepository;
+import com.ryc.api.v2.auth.domain.Admin;
+import com.ryc.api.v2.auth.domain.AdminRepository;
 import com.ryc.api.v2.security.dto.CustomUserDetail;
 
 import lombok.RequiredArgsConstructor;
